follow module spec
