+++ x/nameservice/alias.go
@@ -1,7 +1,8 @@
 package nameservice
 
 import (
-	"github.com/cosmos/sdk-application-tutorial/x/nameservice/types"
+	"github.com/cosmos/sdk-application-tutorial/x/nameservice/internal/keeper"
+	"github.com/cosmos/sdk-application-tutorial/x/nameservice/internal/types"
 )
 
 const (
@@ -11,15 +12,18 @@ const (
 )
 
 var (
+	NewKeeper        = keeper.NewKeeper
+	NewQuerier       = keeper.NewQuerier
 	NewMsgBuyName    = types.NewMsgBuyName
 	NewMsgSetName    = types.NewMsgSetName
 	NewMsgDeleteName = types.NewMsgDeleteName
 	NewWhois         = types.NewWhois
-	ModuleCdc     	 = types.ModuleCdc
-	RegisterCodec 	 = types.RegisterCodec
+	RegisterCodec    = types.RegisterCodec
+	ModuleCdc        = types.ModuleCdc
 )
 
 type (
+	Keeper          = keeper.Keeper
 	MsgSetName      = types.MsgSetName
 	MsgBuyName      = types.MsgBuyName
 	MsgDeleteName   = types.MsgDeleteName
