+++ x/nameservice/internal/keeper/keeper.go
@@ -1,15 +1,15 @@
-package nameservice
+package keeper
 
 import (
 	"github.com/cosmos/cosmos-sdk/codec"
-	"github.com/cosmos/cosmos-sdk/x/bank"
-
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/x/bank"
+	"github.com/cosmos/sdk-application-tutorial/x/nameservice/internal/types"
 )
 
 // Keeper maintains the link to data storage and exposes getter/setter methods for the various parts of the state machine
 type Keeper struct {
-	coinKeeper bank.Keeper
+	CoinKeeper bank.Keeper
 
 	storeKey sdk.StoreKey // Unexposed key to access store from sdk.Context
 
@@ -19,26 +19,26 @@ type Keeper struct {
 // NewKeeper creates new instances of the nameservice Keeper
 func NewKeeper(coinKeeper bank.Keeper, storeKey sdk.StoreKey, cdc *codec.Codec) Keeper {
 	return Keeper{
-		coinKeeper: coinKeeper,
+		CoinKeeper: coinKeeper,
 		storeKey:   storeKey,
 		cdc:        cdc,
 	}
 }
 
 // Gets the entire Whois metadata struct for a name
-func (k Keeper) GetWhois(ctx sdk.Context, name string) Whois {
+func (k Keeper) GetWhois(ctx sdk.Context, name string) types.Whois {
 	store := ctx.KVStore(k.storeKey)
 	if !k.IsNamePresent(ctx, name) {
-		return NewWhois()
+		return types.NewWhois()
 	}
 	bz := store.Get([]byte(name))
-	var whois Whois
+	var whois types.Whois
 	k.cdc.MustUnmarshalBinaryBare(bz, &whois)
 	return whois
 }
 
 // Sets the entire Whois metadata struct for a name
-func (k Keeper) SetWhois(ctx sdk.Context, name string, whois Whois) {
+func (k Keeper) SetWhois(ctx sdk.Context, name string, whois types.Whois) {
 	if whois.Owner.Empty() {
 		return
 	}
@@ -104,4 +104,3 @@ func (k Keeper) IsNamePresent(ctx sdk.Context, name string) bool {
 	store := ctx.KVStore(k.storeKey)
 	return store.Has([]byte(name))
 }
-
